Moved associated methods
