+++ server/src/server_data/endpoints/stock_endpoints.rs
@@ -1,5 +1,5 @@
 use crate::models::stock::NewStock;
-use crate::server_data::endpoints::handle_api_result;
+use crate::server_data::endpoints::ApiError;
 use crate::server_data::repos::Repo;
 use crate::server_data::repos::stock_repo::StockRepo;
 use actix_web::web;
@@ -8,7 +8,7 @@ use actix_web::{get, post, HttpResponse, Result};
 #[get("/stocks")]
 pub async fn get_stocks(repo: web::Data<Repo>) -> Result<HttpResponse> {
     let stocks_result = repo.get_stocks().await;
-    handle_api_result(stocks_result)
+    ApiError::handle(stocks_result)
 }
 
 #[get("/stocks/{id}")]
@@ -17,7 +17,7 @@ pub async fn get_stock(
     id: web::Path<i32>,
 ) -> Result<HttpResponse> {
     let stock_result = repo.get_stock_by_id(*id).await;
-    handle_api_result(stock_result)
+    ApiError::handle(stock_result)
 }
 
 #[post("stocks")]
@@ -32,5 +32,5 @@ pub async fn create_stock(
         share: stock_data.share,
     };
     let stock_result = repo.create_stock(new_stock).await;
-    handle_api_result(stock_result)
+    ApiError::handle(stock_result)
 }
