+++ server/src/server_data/endpoints/stonker_endpoints.rs
@@ -1,6 +1,6 @@
 use crate::models::stonker::NewStonker;
 use crate::repos::stonker_repo::StonkerRepo;
-use crate::server_data::endpoints::handle_api_result;
+use crate::server_data::endpoints::ApiError;
 use crate::server_data::repos::Repo;
 use actix_web::web;
 use actix_web::{get, post, HttpResponse, Result};
@@ -8,7 +8,7 @@ use actix_web::{get, post, HttpResponse, Result};
 #[get("/stonkers")]
 pub async fn get_stonkers(repo: web::Data<Repo>) -> Result<HttpResponse> {
     let stonkers_result = repo.get_stonkers().await;
-    handle_api_result(stonkers_result)
+    ApiError::handle(stonkers_result)
 }
 
 #[get("/stonkers/{id}")]
@@ -17,7 +17,7 @@ pub async fn get_stonker(
     id: web::Path<i32>,
 ) -> Result<HttpResponse> {
     let stonker_result = repo.get_stonker_by_id(*id).await;
-    handle_api_result(stonker_result)
+    ApiError::handle(stonker_result)
 }
 
 #[get("/stonkers/{id}/overview")]
@@ -26,7 +26,7 @@ pub async fn get_stonker_overview(
     id: web::Path<i32>,
 ) -> Result<HttpResponse> {
     let stonker_result = repo.get_stonker_overview(*id).await;
-    handle_api_result(stonker_result)
+    ApiError::handle(stonker_result)
 }
 
 #[post("stonkers")]
@@ -39,7 +39,7 @@ pub async fn create_stonker(
         balance: stonker_data.balance,
     };
     let stonker_result = repo.create_stonker(new_stonker).await;
-    handle_api_result(stonker_result)
+    ApiError::handle(stonker_result)
 }
 
 #[get("/stonkers/{id}/stocks")]
@@ -48,5 +48,5 @@ pub async fn get_stonker_stocks(
     id: web::Path<i32>,
 ) -> Result<HttpResponse> {
     let stonker_stocks = repo.get_stonker_stocks(*id).await;
-    handle_api_result(stonker_stocks)
+    ApiError::handle(stonker_stocks)
 }
