+++ utils/src/datagen/mod.rs
@@ -1,10 +1,10 @@
-mod generator;
-mod index_vec;
-
 pub mod news;
 pub mod stonkers;
 pub mod company;
 pub mod history;
+
+mod generator;
+mod index_vec;
 
 pub use generator::Generator;
 pub use index_vec::IndexVec;
